Fix anr issue when Modify SMS contact
Each time when checkForTooManyRecipients, a list of recipients numbers
will be calculated, while we only need the count, not the specific list.
And too many times calculating lead to anr.

The list of recipients numbers wil be calculated when update recipients.
Just save the size of this list and use it in checkForTooManyRecipients,
will reduce many times of calculating.

Change-Id: I77f5eb5bd1e6c0056a7b1ad2d7442c2faf765c91
CRs-Fixed: 968229