Fix unread SMS can not be found in "Messaging"
When new mesasge comes, a thread id record will be created in thread table
in database if contact is new, then a message record with this thread
id will be created in sms table. On the other hand , when "Messaging"
opens, it will load all threads and check if thread really has messages
, if thread without any message exist, this thread record will be deleted
because such thread is thought as obsolete. If this delete operation
happen between thread id record creation and message record creation when
new message comes, this new message can not be found sine its thread id
not exist.
To avoid incorrect delete, before the delete operation , check if new
coming message infomation is saving to database, if yes, delay the
delete operation by 1000 ms.

Change-Id: If7b6eb3f92bc2af482cf09b27a9807d735724f23
CRs-Fixed: 690984